requirements.txt
pycodestyle
autopep8
yandex-taxi-testsuite